Downgrade open-graph-scraper and undici
There is an issue currently with Next.js 14 and undici, something to do
with SWC and a relatively recent change to undici. In order to fix this
for now, undici needed to be downgraded to a previous major version
before this commit:

  https://github.com/nodejs/undici/commit/c60acb50353e13ab2d14f4be82071028590a085c

This commit added the problematic syntax that SWC couldn't parse.