Fix individual link edit forms
Using a global dialog for all of the individual edit link forms caused
every single link's edit form to be opened whenever you opened the link
form for a link. The last rendered one would win.

These changes go back to individual link state for whether the form is
open, but also uses the `useGlobalForm` hook to "register" the form so
that all the other forms would know that this form is open.